# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CoverageResponse": {
        "properties": {
          "coverage_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Pct"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/SyllabusTopic"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "space_id",
          "coverage_pct",
          "topics",
          "generated_at"
        ],
        "title": "CoverageResponse",
        "type": "object"
      },
      "CreateSpaceRequest": {
        "description": "Payload for creating a Learning Space, e.g. ``{\"name\": \"Claude Code\"}``.",
        "properties": {
          "goal_text": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Text"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateSpaceRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanItem": {
        "properties": {
          "kind": {
            "enum": [
              "overdue_review",
              "uncovered_topic"
            ],
            "title": "Kind",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "review_item_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Item Id"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "title",
          "rationale"
        ],
        "title": "PlanItem",
        "type": "object"
      },
...
      "ReviewItem": {
        "description": "One reviewable interview point, materialized from a source's summary.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Reviewed At"
          },
          "prompt_text": {
            "title": "Prompt Text",
            "type": "string"
          },
          "source_id": {
            "format": "uuid",
            "title": "Source Id",
            "type": "string"
          },
          "source_title": {
            "title": "Source Title",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_id",
          "source_title",
          "space_id",
          "prompt_text",
          "last_reviewed_at"
        ],
        "title": "ReviewItem",
        "type": "object"
      },
      "ReviewQueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total_pending": {
            "title": "Total Pending",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total_pending"
        ],
        "title": "ReviewQueueResponse",
        "type": "object"
      },
      "ReviewRebuildResponse": {
        "description": "Outcome of one bounded backfill batch.",
        "properties": {
          "failed": {
            "title": "Failed",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "processed": {
            "title": "Processed",
            "type": "integer"
          }
        },
        "required": [
          "processed",
          "failed",
          "pending"
        ],
        "title": "ReviewRebuildResponse",
        "type": "object"
      },
...
      "SourceCounts": {
        "description": "Per-type capture counts for a space, as returned by the counts RPC.",
        "properties": {
          "ai_chat": {
            "default": 0,
            "title": "Ai Chat",
            "type": "integer"
          },
          "article": {
            "default": 0,
            "title": "Article",
            "type": "integer"
          },
          "note": {
            "default": 0,
            "title": "Note",
            "type": "integer"
          },
          "pdf": {
            "default": 0,
            "title": "Pdf",
            "type": "integer"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "youtube": {
            "default": 0,
            "title": "Youtube",
            "type": "integer"
          }
        },
        "title": "SourceCounts",
        "type": "object"
      },
...
      "SpaceListResponse": {
        "properties": {
          "spaces": {
            "items": {
              "$ref": "#/components/schemas/SpaceSummary"
            },
            "title": "Spaces",
            "type": "array"
          }
        },
        "required": [
          "spaces"
        ],
        "title": "SpaceListResponse",
        "type": "object"
      },
      "SpaceSummary": {
        "description": "A space plus enough aggregate detail to render a dashboard tile.",
        "properties": {
          "coverage_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Pct"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "goal_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Text"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_captured_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Captured At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "source_counts": {
            "$ref": "#/components/schemas/SourceCounts",
            "default": {
              "ai_chat": 0,
              "article": 0,
              "note": 0,
              "pdf": 0,
              "total": 0,
              "youtube": 0
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ],
        "title": "SpaceSummary",
        "type": "object"
      },
...
      "StudyPlanResponse": {
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            },
            "title": "Items",
            "type": "array"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "items",
          "generated_at"
        ],
        "title": "StudyPlanResponse",
        "type": "object"
      },
...
      "SyllabusTopic": {
        "description": "One inferred syllabus topic and whether captured concepts cover it.",
        "properties": {
          "covered": {
            "title": "Covered",
            "type": "boolean"
          },
          "label": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "covered"
        ],
        "title": "SyllabusTopic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Phase 0 thin capture endpoint: writes raw artifacts to Supabase Storage.",
    "title": "exomemri Capture API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/spaces": {
      "get": {
        "operationId": "list_spaces_v1_spaces_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Spaces",
        "tags": [
          "spaces"
        ]
      },
      "post": {
        "operationId": "create_space_v1_spaces_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Space",
        "tags": [
          "spaces"
        ]
      }
    },
    "/v1/spaces/{space_id}/coverage": {
      "get": {
        "description": "The space's inferred syllabus and how much of it is covered.\n\nGenerated on first read and cached; regenerated automatically once the\nspace's mapped concepts change since the last generation.",
        "operationId": "get_space_coverage_v1_spaces__space_id__coverage_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Space Coverage",
        "tags": [
          "coverage"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/plan": {
      "get": {
        "description": "Overdue reviews and uncovered coverage topics, resequenced into one list.\n\nNo new gap detection and nothing cached \u2014 composed fresh from the review\nqueue and coverage on every call.",
        "operationId": "get_study_plan_v1_spaces__space_id__plan_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study Plan",
        "tags": [
          "plan"
        ]
      }
    },
    "/v1/spaces/{space_id}/review/rebuild": {
      "post": {
        "description": "Generate review items for one bounded batch of unprocessed sources.\n\nNew captures get items automatically once the pipeline's summary stage\nruns; this exists for sources captured before the queue shipped, and for\nretrying ones whose pipeline run failed. Bounded per call, so the client\nloops until ``pending`` is 0.",
        "operationId": "rebuild_review_items_v1_spaces__space_id__review_rebuild_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRebuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Review Items",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/spaces/{space_id}/review/today": {
      "get": {
        "description": "Items due for review today in one space \u2014 the \"study today\" queue.",
        "operationId": "get_today_queue_v1_spaces__space_id__review_today_get",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Today Queue",
        "tags": [
          "review"
        ]
      }
    },
    "/v1/spaces/{space_id}/review/{item_id}/reviewed": {
      "post": {
        "operationId": "mark_item_reviewed_v1_spaces__space_id__review__item_id__reviewed_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Item Reviewed",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}